Wow, check description - kind of working
So, it's a bit of mess. I see why the original author didn't do this, it's a lot of work... But basically, because the bot is stateless, I need to move the state around each interaction. This will work though, so I'm gonna do this a little later, for now I'm gonna rest